+++ lib/src/styles/theme.dart
@@ -177,6 +177,7 @@ class ThemeData with Diagnosticable {
   final Color acrylicBackgroundColor;
   final Color micaBackgroundColor;
   final Color menuColor;
+  final Color cardColor;
 
   final Duration fasterAnimationDuration;
   final Duration fastAnimationDuration;
@@ -251,6 +252,7 @@ class ThemeData with Diagnosticable {
     required this.pillButtonBarTheme,
     required this.bottomSheetTheme,
     required this.menuColor,
+    required this.cardColor,
   });
 
   static ThemeData light() {
@@ -279,6 +281,7 @@ class ThemeData with Diagnosticable {
     Color? checkedColor,
     Color? borderInputColor,
     Color? menuColor,
+    Color? cardColor,
     Duration? fasterAnimationDuration,
     Duration? fastAnimationDuration,
     Duration? mediumAnimationDuration,
@@ -338,6 +341,7 @@ class ThemeData with Diagnosticable {
         ? const Color.fromRGBO(0, 0, 0, 0.4458)
         : const Color.fromRGBO(255, 255, 255, 0.5442);
     menuColor ??= isLight ? const Color(0xFFf9f9f9) : const Color(0xFF2c2c2c);
+    cardColor ??= isLight ? const Color(0xFFf3f3f3) : const Color(0xFF2e2e2e);
     typography = Typography.standard(brightness: brightness)
         .merge(typography)
         .apply(fontFamily: fontFamily);
@@ -420,6 +424,7 @@ class ThemeData with Diagnosticable {
       pillButtonBarTheme: pillButtonBarTheme,
       bottomSheetTheme: bottomSheetTheme,
       menuColor: menuColor,
+      cardColor: cardColor,
     );
   }
 
@@ -444,6 +449,7 @@ class ThemeData with Diagnosticable {
       uncheckedColor: Color.lerp(a.uncheckedColor, b.uncheckedColor, t)!,
       checkedColor: Color.lerp(a.checkedColor, b.checkedColor, t)!,
       borderInputColor: Color.lerp(a.borderInputColor, b.borderInputColor, t)!,
+      cardColor: Color.lerp(a.cardColor, b.cardColor, t)!,
       fasterAnimationDuration:
           lerpDuration(a.fasterAnimationDuration, b.fasterAnimationDuration, t),
       fastAnimationDuration:
@@ -505,6 +511,7 @@ class ThemeData with Diagnosticable {
     Color? checkedColor,
     Color? borderInputColor,
     Color? menuColor,
+    Color? cardColor,
     Duration? fasterAnimationDuration,
     Duration? fastAnimationDuration,
     Duration? mediumAnimationDuration,
@@ -551,6 +558,7 @@ class ThemeData with Diagnosticable {
           acrylicBackgroundColor ?? this.acrylicBackgroundColor,
       micaBackgroundColor: micaBackgroundColor ?? this.micaBackgroundColor,
       menuColor: menuColor ?? this.menuColor,
+      cardColor: cardColor ?? this.cardColor,
       fasterAnimationDuration:
           fasterAnimationDuration ?? this.fasterAnimationDuration,
       fastAnimationDuration:
@@ -597,7 +605,8 @@ class ThemeData with Diagnosticable {
       ..add(ColorProperty('scaffoldBackgroundColor', scaffoldBackgroundColor))
       ..add(ColorProperty('acrylicBackgroundColor', acrylicBackgroundColor))
       ..add(ColorProperty('micaBackgroundColor', micaBackgroundColor))
-      ..add(ColorProperty('menuColor', menuColor));
+      ..add(ColorProperty('menuColor', menuColor))
+      ..add(ColorProperty('cardColor', cardColor));
     properties.add(EnumProperty('brightness', brightness));
     properties.add(DiagnosticsProperty<Duration>(
       'slowAnimationDuration',